Migrate gen_1k_* to Criterion
